dom/alarm/test/test_alarm_non_permitted_app.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Non-Permitted Application for Alarm API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();
if (SpecialPowers.hasPermission("alarms", document)) {
  SpecialPowers.removePermission("alarms", document);
  window.location.reload();
} else {
  SpecialPowers.pushPrefEnv({"set": [["dom.mozAlarms.enabled", true]]}, function() {
    SpecialPowers.removePermission("alarms", document);
    ok(!('mozAlarms' in navigator),
       "navigator.mozAlarms should not exist without permission");
    ok(!('AlarmsManager' in window),
       "Interface AlarmsManager should not exist without permission");
    SpecialPowers.addPermission("alarms", true, document);
    SimpleTest.finish();
  });
}
</script>
</pre>
</body>
</html>
